requirements.txt
groq>=0.37.0
openai>=1.0.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0
questionary>=2.0.0
